Break out terrain & world gen code from world class
